+++ ft_printf/ft_printf_utils.h
@@ -6,7 +6,7 @@
 /*   By: cvillene <cvillene@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/14 08:18:07 by cvillene          #+#    #+#             */
-/*   Updated: 2025/11/14 09:21:32 by cvillene         ###   ########.fr       */
+/*   Updated: 2025/11/14 10:37:32 by cvillene         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@
 # include "Libft/libft.h"
 
 char	*ft_reversestr(char *s);
-char	*ft_dectohex(long long n, int islower);
-int		ft_nbrlen(long n);
+char	*ft_dectohex(unsigned long n, int islower);
+int		ft_nbrlen(unsigned long n);
 char	*ft_ltoa(long n);
 
 #endif
